agent/requirements.txt
browser-use
python-dotenv
langchain_openai
